Renamed issuecolumn package to issuepanel
