+++ src/main/java/ui/issuepanel/FilterPanel.java
@@ -1,4 +1,4 @@
-package ui.issuecolumn;
+package ui.issuepanel;
 
 import backend.interfaces.IModel;
 import backend.resource.TurboIssue;
