+++ src/main/java/ui/listpanel/ListPanel.java
@@ -8,8 +8,8 @@ import javafx.scene.layout.Priority;
 import ui.UI;
 import ui.components.KeyboardShortcuts;
 import ui.components.IssueListView;
-import ui.issuecolumn.PanelControl;
-import ui.issuecolumn.FilterPanel;
+import ui.issuepanel.PanelControl;
+import ui.issuepanel.FilterPanel;
 import util.KeyPress;
 import util.events.IssueSelectedEvent;
 import util.events.testevents.UIComponentFocusEvent;
